% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_diagnostics.R
\name{get_diagnostics}
\alias{get_diagnostics}
\title{Function to extract Stan diagnostics from an HPR model}
\usage{
get_diagnostics(object = NULL, verbose = FALSE)
}
\arguments{
\item{object}{The results object from a run of hpr.}

\item{verbose}{A logical indicator of whether a full cmdstan diagnostic report should
be printed to the console. The default is false.}
}
\value{
A dataframe with columns:
\describe{
   \item{Divergences}{The number of HMC samples that ended in a divergence.}
   \item{Max_Treedepth}{The number of HMC samples that had a max_treedepth warning.}
   \item{Rhat}{The number of f parameters that had Rhat greater than 1.1, using the adjusted Rhat
   of Vehtari et al. (Bayesian Analysis, 2021).}
   \item{Min_Ess_Bulk}{The minimum effective sample size in the bulk of the posterior across
   the f parameters. This estimated according to Vehtari et al. (Bayesian Analysis, 2021).}
   \item{Min_Ess_Tail}{The minimum effective sample size in the tails of the posterior across
   the f parameters. This estimated according to Vehtari et al. (Bayesian Analysis, 2021).}
   \item{Num_Param}{The length of the f vector (the systematic component of the model), which
   is a function of all other parameters in the model.}
   \item{Num_Samples}{The number of HMC posterior samples.}
   \item{Time}{The computing time of Stan sampling.}
}
For more information on these metrics, please see Chase et al. (2022+) or the Stan reference manual.
}
\description{
Function to extract Stan diagnostics from an HPR model
}
\examples{
X <- as.matrix(dat$Day, ncol = 1)
y <- dat$Temperature

mymodel <- hpr(y = y, X = X, family = "gaussian")
get_diagnostics(mymodel)

}
