% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_f.R
\name{get_f}
\alias{get_f}
\title{Function to extract the predicted horseshoe curve fits from an HPR model}
\usage{
get_f(object = NULL, alpha = 0.05)
}
\arguments{
\item{alpha}{}
}
\value{

}
\description{
Function to extract the predicted horseshoe curve fits from an HPR model
}
\examples{

}
