% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_H.R
\name{get_H}
\alias{get_H}
\title{Function to extract the nonlinear component of an HPR model, in the presence
of augmented data}
\usage{
get_H(object = NULL, alpha = 0.05)
}
\arguments{
\item{object}{The results object from a run of hpr.}

\item{alpha}{The uncertainty level for the values of the ho; the
default is 0.05 (which corresponds to 95\% credible intervals).}
}
\value{
A dataframe with N rows (corresponding to the rows of X, X_aug) and columns:
\describe{
   \item{Mean}{the mean of the posterior samples of the horseshoe component}
   \item{Median}{the median of the posterior samples of the horseshoe component}
   \item{Lower}{the alpha/2 percentile of the posterior samples of the horseshoe component}
   \item{Upper}{the 1-alpha/2 percentile of the posterior samples of the horseshoe component}
}
}
\description{
Function to extract the estimated value of the horseshoe component of the model
from an HPR model in the presence of augmented data. If there are no augmented data,
use get_preds. For posterior predictions, use post_predict.
}
\examples{
X <- as.matrix(dat$Day, ncol = 1)
y <- dat$Temperature

mymodel <- hpr(y = y, X = X, family = "gaussian")
my_H <- get_H(mymodel)

}
