% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_preds.R
\name{get_preds}
\alias{get_preds}
\title{Function to extract f, the systematic component of an HPR model}
\usage{
get_preds(object = NULL, alpha = 0.05)
}
\arguments{
\item{object}{The results object from a run of hpr.}

\item{alpha}{The uncertainty level for the values of f; the
default is 0.05 (which corresponds to 95\% credible intervals).}
}
\value{
A dataframe with N rows (corresponding to the rows of X, Z) and columns:
\describe{
   \item{Mean}{the mean of the posterior samples of f}
   \item{Median}{the median of the posterior samples of f}
   \item{Lower}{the alpha/2 percentile of the posterior samples of f}
   \item{Upper}{the 1-alpha/2 percentile of the posterior samples of f}
}
}
\description{
Function to extract the estimated value of f (the systematic component of the model)
from an HPR model. For posterior predictions, use post_predict.
}
\examples{
X <- as.matrix(dat$Day, ncol = 1)
y <- dat$Temperature

mymodel <- hpr(y = y, X = X, family = "gaussian")
my_f <- get_preds(mymodel)
my_f$x <- dat$Day

}
