% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpr.R
\name{hpr}
\alias{hpr}
\title{Function to fit a Bayesian horseshoe process regression}
\usage{
hpr(
  y = NULL,
  X = NULL,
  Z = NULL,
  X_aug = NULL,
  family = NULL,
  monotonic_terms = NULL,
  monotonic_approach = NULL,
  aug_approach = "MCMC",
  seed = NULL,
  chains = 4,
  iter_warmup = 1000,
  iter_sampling = 2000,
  max_treedepth = 12,
  adapt_delta = 0.95,
  c = 0.01,
  verbose = FALSE
)
}
\arguments{
\item{verbose}{}
}
\value{

}
\description{
This function fits a Bayesian horseshoe process regression: it calculates the
posterior distribution for a set of q horseshoe process terms on some continuous
predictors X, and a set of p linear predictors Z. Outcomes can be Gaussian,
binary, or Poisson distributed. Monotonic constraints can be enforced on the
horseshoe process terms, if desired. At present, this model is implemented for
q = 1 or 2, although future versions may permit higher values for q.
If the predictor(s) X are sparsely distributed over their domain, it may be
wise to add some additional augmented data to even the grid of X or to obtain
predictions/interpolations where desired. More information can be found in Chase et al. (2023).
}
\examples{

}
