% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_predict.R
\name{post_predict}
\alias{post_predict}
\title{Function to extract posterior predictions from an HPR model. Note that these are
posterior predictions; if you want estimates of the estimated trajectory f
(the systematic model component), use get_preds.}
\usage{
post_predict(object = NULL, alpha = 0.05)
}
\arguments{
\item{object}{The results object from a run of hpr, for which a new_X and
new_Z matrix were input.}

\item{alpha}{The uncertainty level for posterior prediction intervals; the
default is 0.05 (which corresponds to 95\% prediction intervals).}
}
\value{
A dataframe with N_new rows (corresponding to the values of new_X, new_Z) and columns:
\describe{
   \item{Mean}{the mean of the posterior prediction samples}
   \item{Median}{the median of the posterior prediction samples}
   \item{Lower}{the alpha/2 percentile of the posterior prediction samples}
   \item{Upper}{the 1-alpha/2 percentile of the posterior prediction samples}
}
}
\description{
Function to extract posterior predictions from an HPR model. Note that these are
posterior predictions; if you want estimates of the estimated trajectory f
(the systematic model component), use get_preds.
}
\examples{
X <- as.matrix(dat$Day, ncol = 1)
y <- dat$Temperature

mymodel <- hpr(y = y, X = X, family = "gaussian", new_X = X)
post_preds <- post_predict(mymodel)
post_preds$x <- dat$Day

}
