% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_betas.R
\name{get_betas}
\alias{get_betas}
\title{Function to extract the estimates of the linear coefficients from an HPR model}
\usage{
get_betas(object = NULL, alpha = 0.05, var_names = NULL)
}
\arguments{
\item{var_names}{}
}
\value{

}
\description{
Function to extract the estimates of the linear coefficients from an HPR model
}
\examples{

}
