% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_diagnostics.R
\name{get_diagnostics}
\alias{get_diagnostics}
\title{Function to extract Stan diagnostics from an HPR model}
\usage{
get_diagnostics(object = NULL, verbose = FALSE)
}
\arguments{
\item{object}{}
}
\value{

}
\description{
Function to extract Stan diagnostics from an HPR model
}
\examples{

}
