% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_preds.R
\name{get_preds}
\alias{get_preds}
\title{Function to extract predicted y values from an HPR model}
\usage{
get_preds(object = NULL, alpha = 0.05)
}
\arguments{
\item{alpha}{}
}
\value{

}
\description{
Function to extract predicted y values from an HPR model
}
\examples{

}
